+++ CHANGELOG.md
@@ -6,10 +6,9 @@
 - (#282) Branch support added. You can now specify your branches with a `#branch` fragment at the end of your git url. E.g.: https://github.com/mesg-foundation/service-ethereum-erc20#websocket
 
 #### Added
+- (#242) Add more details in command `mesg-core service validate`
+
 #### Removed
-
-- (#203) Remove useless packages
-
 #### Fixed
 
 ## [v1.0.0]